feat: refacto metric structure
