+++ src/github/metrics/active-contribution-weeks/helpers/remove-duplicate-weeks.ts
@@ -1,0 +1,14 @@
+import { WeekObject } from '../types';
+
+/**
+ * Remove all duplicate weeks.
+ */
+export const removeDuplicateWeeks = (weeks: WeekObject[]): WeekObject[] => {
+  return weeks.filter(
+    (week, index) =>
+      weeks.findIndex(
+        (item) =>
+          item.weekNumber === week.weekNumber && item.year === week.year,
+      ) === index,
+  );
+};
