+++ src/github/metrics/active-contribution-weeks/types/contributions-by-repositories.interface.ts
@@ -1,0 +1,7 @@
+import { RepositoriesQuery } from '@/github/metrics/helpers';
+
+export interface ContributionsByRepository extends RepositoriesQuery {
+  contributions: {
+    nodes?: { occurredAt: string }[];
+  };
+}
