+++ src/github/metrics/active-contribution-weeks/types/created-repositories.interface.ts
@@ -1,0 +1,5 @@
+import { RepositoriesQuery } from '@/github/metrics/helpers';
+
+export interface CreatedRepositories extends RepositoriesQuery {
+  occurredAt: string;
+}
