+++ src/github/metrics/active-contribution-weeks/types/index.ts
@@ -1,0 +1,3 @@
+export * from './contributions-by-repositories.interface';
+export * from './created-repositories.interface';
+export * from './week-object.interface';
