+++ src/github/metrics/active-contribution-weeks/types/week-object.interface.ts
@@ -1,0 +1,4 @@
+export interface WeekObject {
+  year: number;
+  weekNumber: number;
+}
