+++ src/github/metrics/helpers/get-open-source-repositories.ts
@@ -1,0 +1,31 @@
+const openSourceLicenses = ['mit', 'apache-2.0', 'gpl-3.0', 'agpl-3.0'];
+
+/**
+ * Return open source repositories.
+ */
+export function getOpenSourceRepositories<T extends RepositoriesQuery>(
+  contributionsByRepositories: T[],
+): T[] {
+  return contributionsByRepositories.filter(
+    (contributionsByRepository) =>
+      !contributionsByRepository.repository.isFork &&
+      !contributionsByRepository.repository.isPrivate &&
+      openSourceLicenses.includes(
+        contributionsByRepository.repository.licenseInfo?.key,
+      ),
+  );
+}
+
+export interface RepositoriesQuery {
+  repository: RepositoryQuery;
+}
+
+interface RepositoryQuery {
+  licenseInfo?: LicenseQuery;
+  isFork: boolean;
+  isPrivate: boolean;
+}
+
+interface LicenseQuery {
+  key: string;
+}
