+++ src/github/metrics/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './get-open-source-repositories';
