+++ src/github/metrics/metric-name.ts
@@ -3,4 +3,5 @@ export enum MetricName {
   issuePullRequestRatio = 'issuePullRequestRatio',
   maintainedRepositoryCount = 'maintainedRepositoryCount',
   totalContributions = 'totalContributions',
+  activeContributionWeeks = 'activeContributionWeeks',
 }
